Сделал задание 5
